Introduce allTests for all SILTests test cases excluding InstructionTests
