+++ Tests/LinuxMain.swift
@@ -1,9 +1,18 @@
 import QuoteTests
+import SILTests
 import XCTest
 
 let tests = [
-    testCase(DescriptionTests.allTests),
-    testCase(QuoteTests.allTests),
+    // QuoteTests
+    testCase(CompilationTests.allTests),
+    testCase(QuoteTests.DescriptionTests.allTests),
     testCase(StructureTests.allTests),
+    // SILTests
+    testCase(BitcodeTests.allTests),
+    testCase(BitsTests.allTests),
+    testCase(BitstreamTests.allTests),
+    testCase(SILTests.DescriptionTests.allTests),
+    testCase(ModuleTests.allTests),
+    testCase(PrinterTests.allTests),
 ]
 XCTMain(tests)
